+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 - Timeline new structure
 - Added classes for margin & padding - 2, 4, 8, 16, 20 & 32
+- Fix hover & disabled states